@@ -4,25 +4,26 @@ description: A new Flutter project.
 version: 3.0.0
 
 environment:
-  sdk: '>=3.3.4 <4.0.0'
+  sdk: '>=3.7.2 <4.0.0'
 
 dependencies:
   animations: ^2.0.11
   cupertino_icons: ^1.0.8
   flutter:
     sdk: flutter
-  font_awesome_flutter: ^10.7.0
-  intl: ^0.19.0
+  flutter_riverpod: ^2.6.1
+  font_awesome_flutter: ^10.8.0
+  intl: ^0.20.2
   provider: ^6.1.2
-  shared_preferences: ^2.2.3
+  shared_preferences: ^2.5.2
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.13.1
-  flutter_lints: ^3.0.2
-  flutter_native_splash: ^2.4.0
+  flutter_launcher_icons: ^0.14.3
+  flutter_lints: ^5.0.0
+  flutter_native_splash: ^2.4.5
   flutter_test:
     sdk: flutter
-    
+
 flutter_launcher_icons:
   android: true
   ios: true
