@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:provider/provider.dart';
-import 'package:veggie_android_seasons/data/app_state.dart';
-import 'package:veggie_android_seasons/data/veggie_preferences.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:veggie_android_seasons/screens/home_screen.dart';
 import 'package:veggie_android_seasons/veggie_styles.dart';
 
@@ -10,29 +8,22 @@ void main() {
   WidgetsFlutterBinding.ensureInitialized();
   SystemChrome.setPreferredOrientations(
       [DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]);
-  runApp(MyApp());
+  runApp(ProviderScope(child: MyApp()));
 }
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        ChangeNotifierProvider.value(value: AppState()),
-        ChangeNotifierProvider(create: (_) => VeggiePrefs()..load())
-      ],
-      child: MaterialApp(
-        title: 'Veggie Android Seasons',
-        theme: ThemeData(
-          useMaterial3: true,
-          //colorSchemeSeed: Colors.red,
-          colorScheme: ColorScheme.fromSwatch(primarySwatch: Colors.red),
-          //primarySwatch: Colors.red,
-          scaffoldBackgroundColor: VeggieStyles.scaffoldBackground,
-          brightness: Brightness.light,
-        ),
-        home: HomeScreen(),
+    return MaterialApp(
+      title: 'Veggie Android Seasons',
+      theme: ThemeData(
+        //colorSchemeSeed: Colors.red,
+        colorScheme: ColorScheme.fromSwatch(primarySwatch: Colors.red),
+        //primarySwatch: Colors.red,
+        scaffoldBackgroundColor: VeggieStyles.scaffoldBackground,
+        brightness: Brightness.light,
       ),
+      home: HomeScreen(),
     );
   }
 }
