@@ -1,22 +1,22 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
-import 'package:provider/provider.dart';
-import 'package:veggie_android_seasons/data/app_state.dart';
+import 'package:veggie_android_seasons/data/app_notifier.dart';
 import 'package:veggie_android_seasons/data/veggie.dart';
-import 'package:veggie_android_seasons/data/veggie_preferences.dart';
+import 'package:veggie_android_seasons/data/veggie_preferences_notifier.dart';
 import 'package:veggie_android_seasons/veggie_styles.dart';
 import 'package:veggie_android_seasons/widgets/veggie_card.dart';
 
-class VeggieListScreen extends StatelessWidget {
+class VeggieListScreen extends ConsumerWidget {
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     var dateString = DateFormat('MMMM y').format(DateTime.now());
-    final appState = Provider.of<AppState>(context);
-    final prefs = Provider.of<VeggiePrefs>(context);
+    final appState = ref.watch(appNotifierProvider);
+
     return DecoratedBox(
       decoration: const BoxDecoration(color: Color(0xffffffff)),
       child: ListView.builder(
-        itemCount: appState.allVeggies.length + 2,
+        itemCount: appState.length + 2,
         itemBuilder: (BuildContext context, int index) {
           if (index == 0) {
             return Padding(
@@ -30,18 +30,20 @@ class VeggieListScreen extends StatelessWidget {
                 ],
               ),
             );
-          } else if (index <= appState.availableVeggies.length) {
-            return _generateVeggieRow(
-                appState.availableVeggies[index - 1], prefs);
-          } else if (index <= appState.availableVeggies.length + 1) {
+          } else if (index <= appState.length) {
+            return _generateVeggieRow(appState[index - 1], ref);
+          } else if (index <= appState.length + 1) {
             return const Padding(
               padding: EdgeInsets.fromLTRB(16, 24, 16, 16),
               child: Text('Not in Season', style: VeggieStyles.headlineText),
             );
           } else {
-            var relativeIndex = index - (appState.availableVeggies.length + 2);
+            final relativeIndex = index - (appState.length + 2);
             return _generateVeggieRow(
-                appState.unavailableVeggies[relativeIndex], prefs,
+                ref
+                    .watch(appNotifierProvider.notifier)
+                    .unavailableVeggies[relativeIndex],
+                ref,
                 inSeason: false);
           }
         },
@@ -49,20 +51,15 @@ class VeggieListScreen extends StatelessWidget {
     );
   }
 
-  Widget _generateVeggieRow(Veggie veggie, VeggiePrefs veggiePrefs,
+  Widget _generateVeggieRow(Veggie veggie, WidgetRef ref,
       {bool inSeason = true}) {
+    final preferredCategories = ref.watch(preferredCategoriesProvider);
     return Padding(
       padding: const EdgeInsets.only(left: 16, right: 16, bottom: 24),
-      child: FutureBuilder<Set<VeggieCategory>>(
-        future: veggiePrefs.preferredCategories,
-        builder: (BuildContext context, AsyncSnapshot snapshot) {
-          final data = snapshot.data ?? <VeggieCategory>{};
-          return VeggieCard(
-              veggie: veggie,
-              isInSeason: inSeason,
-              isPreferredCategory: data.contains(veggie.category));
-        },
-      ),
+      child: VeggieCard(
+          veggie: veggie,
+          isInSeason: inSeason,
+          isPreferredCategory: preferredCategories.contains(veggie.category)),
     );
   }
 }
